LatinIME: correctly handle jni fallbacks
The way it is done till now does not make really sense. If
the gesture lib was flashed with gapps from the user it still returns
the value config_gesture_input_enabled_by_build_config back which returns
false by default on aosp

So do not rely on config_gesture_input_enabled_by_build_config at all and only
decide on the type of lib which was loaded.

Change-Id: Ib785de3dff968e2179972628233c5bf65bbd31df